requirements.txt
fastapi==0.70.0
async-timeout<4.0,>=3.0
stor-blockchain==1.0.4
uvicorn[standard]==0.15.0
aioredis==1.3.0
logzero
aiocache